gpu/capture: lift the capture-format vocabulary into an internal header
Step one of splitting gpu_capture.cpp (6515 lines). This is the semantic half and it is
deliberately its own commit: 60 regions change linkage from internal to external so a
second translation unit can see them. No behaviour changes.

What moves is the file format's shared vocabulary -- the magic/version/limit constants,
the resource-footprint and blob-size helpers, the seed and shader-version helpers, and the
validators -- computed as the closure of what the serialization code calls, not chosen by
hand. split_file.py named the 33 definitions that a serialize/deserialize split would
strand, and promote_internal.py closed that set under its own callees to 60.

gpu_capture.cpp 6515 -> 5183 lines; gpu_capture_internal.hpp 1399. Builds clean.

40 definitions gain `inline`, which is required once more than one translation unit
includes the header.
